src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/prerender-insert-after-stop-expected.txt

CONSOLE ERROR: line 4: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 4: Uncaught TypeError: document.registerElement is not a function
Test that Custom Element should be defined while navigating between frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS customBody is defined.
PASS customBody.toLocaleString() is "[object HTMLBodyElement]"
PASS successfullyParsed is true

TEST COMPLETE
